requirements.txt
altair==4.1.0
backcall==0.2.0
base58==2.1.0
gitdb==4.0.7
GitPython==3.1.18
matplotlib==3.3.3
nltk==3.6.2
notebook==6.1.4
numpy==1.19.3
packaging==20.4
pandas==1.2.4
pickleshare==0.7.5
Pillow==8.0.1
regex==2021.8.3
requests==2.26.0
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
sklearn==0.0
streamlit==0.86.0
textblob==0.15.3